Created Modal component. Refactor AddPost component to be in a Modal. Added Edit and Delete actions. Added animations with Vue.js Transitions. Added responsive styles.

+++ src/store/modules/posts.js
@@ -34,11 +34,13 @@ const actions = {
     }
   },
   async deletePost({ commit }, id) {
-    try {
-      await api.delete(`/posts/${id}`);
-      commit('DELETE_POST', id);
-    } catch (error) {
-      console.log(error);
+    if (confirm('Are you sure?')) {
+      try {
+        await api.delete(`/posts/${id}`);
+        commit('DELETE_POST', id);
+      } catch (error) {
+        console.log(error);
+      }
     }
   }
 };
